requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2
pytest
httpx
python-dotenv